Monotone-Parent: f49f7655bf5cf28a4146c61d6f0897052be0f21b Monotone-Revision: 2d5e6477bc1967b51bbbe812372dc78cb0f8f69f
Monotone-Author: wsourdeau@inverse.ca
Monotone-Date: 2011-07-29T14:16:53
Monotone-Branch: ca.inverse.sogo

+++ OpenChange/plreader.m
@@ -81,7 +81,7 @@ const char *indentationStep = "  ";
   max = [keys count];
 
   [self _outputIndentation: anInt];
-  printf ("{ (%ld) items\n", max);
+  printf ("{ (%ld) items\n", (long) max);
 
   subIndent = anInt + 1;
 
@@ -104,7 +104,7 @@ const char *indentationStep = "  ";
       withIndentation: (NSInteger) anInt
 {
   [self _outputIndentation: anInt];
-  printf ("%lu =\n", count);
+  printf ("%lu =\n", (unsigned long) count);
 }
 
 - (void) displayWithIndentation: (NSInteger) anInt
@@ -115,7 +115,7 @@ const char *indentationStep = "  ";
   max = [self count];
 
   [self _outputIndentation: anInt];
-  printf ("[ (%ld) items\n", max);
+  printf ("[ (%ld) items\n", (long) max);
 
   subIndent = anInt + 1;
